requirements.txt
streamlit
ultralytics
streamlit_toggle
opencv-python
Pillow
torch
lightning
torchvision